Added a small Dependency Injection Container called Dependency Manager.
It supports both shared service of which there must only be one instance at any
time and services that can be recreated frequently. Every managed class is
required to provide a static createService method which can then pass the
dependencies to the constructor.

It makes heavy use of templates, a service is retrieved using
    manager->getService<ServiceName>()

To register a class as a service it needs to create a RegisterService instance
with the appropriate template parameters statically.